chore: move default value set to plugin handler
